Refactor similarity related function
